refactor: Consolidate powers and items filters
