+++ src/routes/(main)/(talents)/powers/+page.ts
